+++ tests/test_reader.py
@@ -632,7 +632,8 @@ def test_add_remove_get_feeds(reader, feed_arg):
 
 
 def test_get_feeds_order_title(reader):
-    """Feeds should be sorted by (with decreasing priority):
+    """When order_by='title', feeds should be sorted by (with decreasing
+    priority):
 
     * feed user_title or feed title; feeds that have neither should appear first
     * feed URL
@@ -667,7 +668,8 @@ def test_get_feeds_order_title(reader):
 
 
 def test_get_feeds_order_title_case_insensitive(reader):
-    """Feeds should be sorted by title in a case insensitive way.
+    """When order_by='title', feeds should be sorted by title in a case
+    insensitive way.
 
     https://github.com/lemon24/reader/issues/103
 
@@ -686,6 +688,40 @@ def test_get_feeds_order_title_case_insensitive(reader):
     reader.update_feeds()
 
     assert list(reader.get_feeds()) == [feed1, feed3, feed2]
+
+
+def test_get_feeds_order_added(reader):
+    """When order_by='added', feeds should be sorted by (with decreasing
+    priority):
+
+    * feed added, descending (feeds that don't have it should appear last)
+    * feed URL
+
+    https://github.com/lemon24/reader/issues/98
+
+    """
+
+    parser = Parser()
+    reader._parse = parser
+
+    reader._now = lambda: datetime(2010, 1, 1)
+    feed1 = parser.feed(1, datetime(2010, 1, 2))
+    reader.add_feed(feed1.url)
+
+    reader._now = lambda: datetime(2010, 1, 2)
+    feed2 = parser.feed(2, datetime(2010, 1, 1))
+    reader.add_feed(feed2.url)
+
+    reader._now = lambda: None
+    feed3 = parser.feed(3, datetime(2010, 1, 3))
+    reader.add_feed(feed3.url)
+
+    assert list(reader.get_feeds(order_by='added')) == [
+        Feed(f.url, None, None, None, None) for f in [feed2, feed1, feed3]]
+
+    reader.update_feeds()
+
+    assert list(reader.get_feeds(order_by='added')) == [feed2, feed1, feed3]
 
 
 def test_set_feed_user_title(reader, feed_arg):